+++ packages/moonstone/internal/Picker/PickerButton.js
@@ -88,7 +88,15 @@ const PickerButtonBase = kind({
 			);
 		} else {
 			return (
-				<IconButton {...rest} backgroundOpacity="transparent" disabled={disabled} small>
+				<IconButton
+					{...rest}
+					backgroundOpacity="transparent"
+					disabled={disabled}
+					holdConfig={{events: [
+						{name: 'hold', time: 800}
+					]}}
+					small
+				>
 					{icon}
 				</IconButton>
 			);