+++ packages/moonstone/CHANGELOG.md
@@ -7,6 +7,7 @@ The following is a curated list of changes in the Enact moonstone module, newest
 ### Fixed
 
 - `moonstone/VideoPlayer` to round the time displayed down to nearest second
+- `moonstone/Picker` to correctly update its value once when button is selected
 
 ## [2.0.2] - 2018-13-01